Added null check when deleting playlist in Admin Settings, so app doesn't crash from trying to delete a deleted playlist

+++ app/src/main/java/com/example/project2_sound_wave/Admin_Settings_Page.java
@@ -176,11 +176,14 @@ public class Admin_Settings_Page extends AppCompatActivity {
             public void onClick(DialogInterface dialog, int which) {
                 repository.delete(user);
                 repository.getUserByUserName(user.getUsername()).removeObserver(userObserver);
+
                 userObserver = null;
                 repository.getPlaylistByUserName(user.getUsername()).observe(Admin_Settings_Page.this, new Observer<Playlist>() {
                     @Override
                     public void onChanged(Playlist playlist) {
-                        repository.delete(playlist);
+                        if (playlist != null) {
+                            repository.delete(playlist);
+                        }
                     }
                 });
                 toastMaker("User successfully deleted");
